# src/RATools.Api/openapi.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RATools.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/filesystem/directories": {
      "get": {
        "tags": [
          "Filesystem"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryBrowseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryBrowseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryBrowseResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/filesystem/resolve-directory": {
      "post": {
        "tags": [
          "Filesystem"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDirectoryRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDirectoryRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDirectoryRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryResolutionResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryResolutionResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryResolutionResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/publish-jobs/{id}": {
      "get": {
        "tags": [
          "PublishJobs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJobDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJobDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJobDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/publish-jobs/{id}/report": {
      "get": {
        "tags": [
          "PublishJobs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublishExecutionReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishExecutionReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishExecutionReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/publish-jobs/{id}/artifacts": {
      "get": {
        "tags": [
          "PublishJobs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublishArtifactsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishArtifactsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishArtifactsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/publish-jobs/{id}/artifacts/{name}/download": {
      "get": {
        "tags": [
          "PublishJobs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/publish-jobs/execute": {
      "post": {
        "tags": [
          "PublishJobs"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePublishJobRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePublishJobRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePublishJobRequestBody"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJobDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJobDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishJobDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/validation/sequence": {
      "post": {
        "tags": [
          "Validation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateSequenceRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateSequenceRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateSequenceRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReportDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/validation/publish-readiness": {
      "post": {
        "tags": [
          "Validation"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishReadinessRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishReadinessRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PublishReadinessRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReadinessReportDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReadinessReportDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReadinessReportDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreatePublishJobRequestBody": {
        "required": [
          "applicationId",
          "sequenceNumber"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "minLength": 1,
            "pattern": "^\\d{4}$",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "DirectoryBrowseEntry": {
        "required": [
          "canBrowse",
          "fullPath",
          "hasChildren",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "fullPath": {
            "type": "string"
          },
          "canBrowse": {
            "type": "boolean"
          },
          "hasChildren": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "DirectoryBrowseResult": {
        "required": [
          "currentPath",
          "directories",
          "parentPath"
        ],
        "type": "object",
        "properties": {
          "currentPath": {
            "type": "string",
            "nullable": true
          },
          "parentPath": {
            "type": "string",
            "nullable": true
          },
          "directories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirectoryBrowseEntry"
            }
          }
        },
        "additionalProperties": false
      },
      "DirectoryResolutionResult": {
        "required": [
          "exists",
          "fullPath",
          "isAccessible",
          "isDirectory"
        ],
        "type": "object",
        "properties": {
          "fullPath": {
            "type": "string"
          },
          "exists": {
            "type": "boolean"
          },
          "isDirectory": {
            "type": "boolean"
          },
          "isAccessible": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "PublishArtifactDto": {
        "required": [
          "contentType",
          "exists",
          "name",
          "path",
          "sizeBytes",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "exists": {
            "type": "boolean"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "PublishArtifactSummaryDto": {
        "required": [
          "fileCount",
          "packageSizeBytes",
          "totalSizeBytes"
        ],
        "type": "object",
        "properties": {
          "fileCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalSizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "packageSizeBytes": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
      "PublishArtifactsDto": {
        "required": [
          "applicationId",
          "artifacts",
          "publishJobId",
          "sequenceNumber"
        ],
        "type": "object",
        "properties": {
          "publishJobId": {
            "type": "string",
            "format": "uuid"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "type": "string"
          },
          "artifacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublishArtifactDto"
            }
          }
        },
        "additionalProperties": false
      },
      "PublishAuditSummaryDto": {
        "required": [
          "latestPublishJobAction",
          "latestPublishJobEventUtc",
          "publishJobEventCount",
          "validationEventCount"
        ],
        "type": "object",
        "properties": {
          "publishJobEventCount": {
            "type": "integer",
            "format": "int32"
          },
          "validationEventCount": {
            "type": "integer",
            "format": "int32"
          },
          "latestPublishJobAction": {
            "type": "string",
            "nullable": true
          },
          "latestPublishJobEventUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PublishExecutionReportDto": {
        "required": [
          "applicationId",
          "artifactSummary",
          "auditSummary",
          "durationMs",
          "errorCount",
          "integrityEvidence",
          "integritySummary",
          "message",
          "publishJob",
          "publishReadiness",
          "reportPath",
          "reportVersion",
          "sequenceNumber",
          "succeeded",
          "validationProfile",
          "validationReport",
          "warningCount",
          "warningSummary"
        ],
        "type": "object",
        "properties": {
          "reportVersion": {
            "type": "string"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "type": "string"
          },
          "validationProfile": {
            "type": "string"
          },
          "reportPath": {
            "type": "string",
            "nullable": true
          },
          "validationReport": {
            "$ref": "#/components/schemas/ValidationReportDto"
          },
          "publishJob": {
            "$ref": "#/components/schemas/PublishJobDto"
          },
          "durationMs": {
            "type": "integer",
            "format": "int64"
          },
          "integritySummary": {
            "$ref": "#/components/schemas/PublishIntegritySummaryDto"
          },
          "integrityEvidence": {
            "$ref": "#/components/schemas/PublishIntegrityEvidenceDto"
          },
          "publishReadiness": {
            "$ref": "#/components/schemas/PublishReadinessReportDto"
          },
          "artifactSummary": {
            "$ref": "#/components/schemas/PublishArtifactSummaryDto"
          },
          "auditSummary": {
            "$ref": "#/components/schemas/PublishAuditSummaryDto"
          },
          "errorCount": {
            "type": "integer",
            "format": "int32"
          },
          "warningCount": {
            "type": "integer",
            "format": "int32"
          },
          "warningSummary": {
            "type": "string",
            "nullable": true
          },
          "succeeded": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PublishIntegrityEvidenceDto": {
        "required": [
          "artifacts",
          "findings"
        ],
        "type": "object",
        "properties": {
          "artifacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublishArtifactEvidenceDto"
            }
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublishIntegrityFindingDto"
            }
          }
        },
        "additionalProperties": false
      },
      "PublishIntegrityFindingDto": {
        "required": [
          "message",
          "path",
          "severity",
          "type"
        ],
        "type": "object",
        "properties": {
          "severity": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PublishIntegritySummaryDto": {
        "required": [
          "isConsistent",
          "mismatchedArtifactsCount",
          "missingFilesCount",
          "missingZipEntriesCount"
        ],
        "type": "object",
        "properties": {
          "isConsistent": {
            "type": "boolean"
          },
          "missingFilesCount": {
            "type": "integer",
            "format": "int32"
          },
          "missingZipEntriesCount": {
            "type": "integer",
            "format": "int32"
          },
          "mismatchedArtifactsCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PublishJobDto": {
        "required": [
          "applicationId",
          "attemptCount",
          "completedUtc",
          "createdUtc",
          "failureReason",
          "id",
          "idempotencyKey",
          "nextAttemptUtc",
          "outputPath",
          "packagePath",
          "sequenceNumber",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "outputPath": {
            "type": "string",
            "nullable": true
          },
          "packagePath": {
            "type": "string",
            "nullable": true
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          },
          "completedUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "failureReason": {
            "type": "string",
            "nullable": true
          },
          "idempotencyKey": {
            "type": "string",
            "nullable": true
          },
          "attemptCount": {
            "type": "integer",
            "format": "int32"
          },
          "nextAttemptUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PublishReadinessReportDto": {
        "required": [
          "applicationId",
          "blockingErrorCount",
          "categorySummaries",
          "findings",
          "isReady",
          "missingMetadataFields",
          "sequenceNumber",
          "status",
          "validationReport",
          "warningCount"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "type": "string"
          },
          "isReady": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "blockingErrorCount": {
            "type": "integer",
            "format": "int32"
          },
          "warningCount": {
            "type": "integer",
            "format": "int32"
          },
          "validationReport": {
            "$ref": "#/components/schemas/ValidationReportDto"
          },
          "missingMetadataFields": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "categorySummaries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublishReadinessCategorySummaryDto"
            }
          },
          "findings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublishReadinessFindingDto"
            }
          }
        },
        "additionalProperties": false
      },
      "PublishReadinessRequestBody": {
        "required": [
          "applicationId",
          "sequenceNumber"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "minLength": 1,
            "pattern": "^\\d{4}$",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ResolveDirectoryRequestBody": {
        "required": [
          "path"
        ],
        "type": "object",
        "properties": {
          "path": {
            "maxLength": 4096,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ValidateSequenceRequestBody": {
        "required": [
          "applicationId",
          "sequenceNumber"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "minLength": 1,
            "pattern": "^\\d{4}$",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ValidationIssueDto": {
        "required": [
          "code",
          "documentId",
          "message",
          "placementId",
          "sectionPath",
          "severity"
        ],
        "type": "object",
        "properties": {
          "severity": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "sectionPath": {
            "type": "string",
            "nullable": true
          },
          "documentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "placementId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ValidationLifecycleMatchDto": {
        "required": [
          "attemptedStrategies",
          "ctdSection",
          "documentId",
          "historicalFinalState",
          "historicalMatchCount",
          "historicalPlacementIds",
          "historicalSequenceNumbers",
          "matchStrategy",
          "operation",
          "resultCode",
          "sequenceNumber"
        ],
        "type": "object",
        "properties": {
          "operation": {
            "type": "string"
          },
          "sequenceNumber": {
            "type": "string"
          },
          "ctdSection": {
            "type": "string"
          },
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "resultCode": {
            "type": "string"
          },
          "matchStrategy": {
            "type": "string"
          },
          "attemptedStrategies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "historicalMatchCount": {
            "type": "integer",
            "format": "int32"
          },
          "historicalSequenceNumbers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "historicalPlacementIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "historicalFinalState": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ValidationReportDto": {
        "required": [
          "applicationId",
          "issues",
          "isValid",
          "lifecycleMatches",
          "sectionMatches",
          "sequenceNumber",
          "validationProfile"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "type": "string"
          },
          "validationProfile": {
            "type": "string"
          },
          "isValid": {
            "type": "boolean"
          },
          "issues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssueDto"
            }
          },
          "sectionMatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationSectionMatchDto"
            }
          },
          "lifecycleMatches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationLifecycleMatchDto"
            }
          }
        },
        "additionalProperties": false
      },
      "ValidationSectionMatchDto": {
        "required": [
          "isStandard",
          "isValid",
          "matchedPrefix",
          "reason",
          "sectionPath"
        ],
        "type": "object",
        "properties": {
          "sectionPath": {
            "type": "string"
          },
          "isValid": {
            "type": "boolean"
          },
          "isStandard": {
            "type": "boolean"
          },
          "matchedPrefix": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}